Chnaged drivers to support SDK3, added driver ZB_NGPB_SWITCH_1
