+++ drivers/ZB_NHPB_SWITCH_1/device.js
@@ -2,10 +2,10 @@
 
 const WiserGenericDevice = require('../../lib/ZB_WiserGenericDevice');
 
-class ZB_PUCK_SWITCH_1 extends WiserGenericDevice {
+class ZB_NHPB_SWITCH_1 extends WiserGenericDevice {
 
   wiserEndpoint = 1;
 
 }
 
-module.exports = ZB_PUCK_SWITCH_1;
+module.exports = ZB_NHPB_SWITCH_1;
