+++ package.json
@@ -3,13 +3,13 @@
   "version": "0.2.0",
   "main": "app.js",
   "devDependencies": {
-    "eslint": "^7.3.1",
+    "eslint": "^7.4.0",
     "eslint-config-athom": "^2.1.0",
     "homey": "^2.5.1"
   },
   "dependencies": {
-    "homey-zigbeedriver": "^1.0.1",
-    "zigbee-clusters": "^1.1.2"
+    "homey-zigbeedriver": "^1.1.0",
+    "zigbee-clusters": "^1.2.0"
   },
   "engines": {
     "node": ">=12.16.1"
